+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.9 - 2018-12-06
 
+* Fix styling on DurationInput
 * Fix Wrong Map schema for departments
 
 ## 0.6.8 - 2018-11-29